+++ instrumentation/opentelemetry-instrumentation-redis/src/opentelemetry/instrumentation/redis/__init__.py
@@ -131,11 +131,15 @@ _REDIS_ASYNCIO_CLUSTER_VERSION = (4, 3, 2)
 _FIELD_TYPES = ["NUMERIC", "TEXT", "GEO", "TAG", "VECTOR"]
 
 
-def _set_connection_attributes(span, conn):
-    if not span.is_recording() or not hasattr(conn, "connection_pool"):
+def _set_connection_attributes(span, instance):
+    if hasattr(instance, "nodes_manager") and hasattr(
+        instance.nodes_manager.default_node, "redis_connection"
+    ):
+        instance = instance.nodes_manager.default_node.redis_connection
+    if not span.is_recording() or not hasattr(instance, "connection_pool"):
         return
     for key, value in _extract_conn_attributes(
-        conn.connection_pool.connection_kwargs
+        instance.connection_pool.connection_kwargs
     ).items():
         span.set_attribute(key, value)